requirements.txt
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
google-api-python-client==1.6.4
google-auth==1.2.0
google-auth-httplib2==0.0.2
google-auth-oauthlib==0.1.1
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
oauth2client==4.1.2
oauthlib==2.0.6
pyasn1==0.3.7
pyasn1-modules==0.1.5
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.12.2
jsonpickle==0.9.5
